requirements.txt
pyyaml>=5.1
openai>=1.0.0
requests>=2.25.1
ansible>=2.9.0
python-nmap>=0.7.1